Complete Three.js refactor
