+++ src/lib/sections/AgoraSpaceDao/AgoraSpaceDao.svelte
@@ -1,5 +1,3 @@
-<section id="agora-space-dao" class="relative bg-agora-gray ">
-  <div class="container px-8 pt-16 xl:pt-24 text-center">
-    <!-- TODO -->
-  </div>
+<section id="agora-space-dao" class="relative h-screen bg-agora-gray text-center">
+  <canvas class="absolute inset-0 w-full h-full" />
 </section>
