+++ src/lib/threejs/noise.js
